+++ src/main/java/eu/olli/cowmoonication/listener/PlayerListener.java
@@ -1,6 +1,7 @@
 package eu.olli.cowmoonication.listener;
 
 import eu.olli.cowmoonication.Cowmoonication;
+import eu.olli.cowmoonication.util.TickDelay;
 import net.minecraftforge.fml.common.eventhandler.SubscribeEvent;
 import net.minecraftforge.fml.common.network.FMLNetworkEvent;
 
@@ -14,5 +15,11 @@ public class PlayerListener {
     @SubscribeEvent
     public void onServerJoin(FMLNetworkEvent.ClientConnectedToServerEvent e) {
         main.getVersionChecker().runUpdateCheck(false);
+        new TickDelay(() -> main.getChatHelper().sendOfflineMessages(), 6 * 20);
+    }
+
+    @SubscribeEvent
+    public void onServerLeave(FMLNetworkEvent.ClientDisconnectionFromServerEvent e) {
+        main.getFriends().saveBestFriends();
     }
 }